Fix ODS pool op and add user op ODS generating to CMake (#6928)
* add adaptive base op and test

* format

* add oneflow-gen-ods

* refine

* fix oneflow-gen-obs adaptive op no group

* fix

* fix bug

* delete useless code

* ad auto gen

* annotation fused dropout and bias pattern

* fix adaptive gen ods bug

* refine

* rm gen

* shield command output ods code

* shield command output ods code

* refine

* refine

* refine

* refine

* refine

* refine

* refine

* fix

* rm

* refine

* refine

* fix

* refien
