+++ oneflow/ir/include/OneFlow/OneFlowOpGetGen.td
@@ -1,4 +1,4 @@
-include "OneFlowDialect.td"
-include "OneFlowEnums.td"
+include "OneFlow/OneFlowDialect.td"
+include "OneFlow/OneFlowEnums.td"
 include "mlir/Interfaces/SideEffectInterfaces.td"
-include "OneFlowBase.td"
+include "OneFlow/OneFlowBase.td"
