+++ oneflow/ir/include/OneFlow/OneFlowPatterns.td
@@ -2,7 +2,7 @@
 #ifndef ONEFLOW_PATTERNS
 #define ONEFLOW_PATTERNS
 
-include "OneFlowOps.td"
+include "OneFlow/OneFlowOps.td"
 
 def IsNotNestedInJit: Constraint<CPred<"(!$0.getDefiningOp()->getParentOfType<FuncOp>()->hasAttr(\"llvm.emit_c_interface\"))">, "">;
 def OutlineMulCast : NativeCodeCall<"OutlineMulCast($_builder, $0, $1)">;
@@ -129,41 +129,6 @@ def FusedScaleTrilPattern2 : Pat<
     (IsGPU $tril_device_tag),
     (IsGPU $scale_device_tag)
   ]
->;
-
-def FusedBiasAddDropoutPattern : Pat<
-  (
-    OneFlow_DropoutOp
-    (
-      OneFlow_BiasAddOp
-        $a,
-        $b,
-        $bias_add_op_name,
-        $bias_add_device_tag,
-        $bias_add_device_name,
-        $bias_add_scope_symbol_id,
-        $bias_add_hierarchy,
-        $axis
-    ),
-    $mask,
-    $_add_to_output,
-    $dropout_op_name,
-    $dropout_device_tag,
-    $dropout_device_name,
-    $dropout_scope_symbol_id,
-    $dropout_hierarchy,
-    $scale
-  ),
-  (OneFlow_FusedBiasAddMaskScaleOp $a, $b, $mask, (GetFirstValue $_add_to_output),
-    $dropout_op_name,
-    $dropout_device_tag,
-    $dropout_device_name,
-    $dropout_scope_symbol_id,
-    $dropout_hierarchy,
-    $axis,
-    $scale
-  ),
-  []
 >;
 
 def FusedBiasAddGeluPattern : Pat<
