+++ oneflow/ir/lib/OneFlow/CMakeLists.txt
@@ -32,6 +32,7 @@ add_mlir_dialect_library(MLIROneFlow
   DEPENDS
   MLIROneFlowOpsIncGen
   prepare_oneflow_third_party
+  oneflow-gen-ods
 
   LINK_LIBS PUBLIC
   ${dialect_libs}
